fix issue

+++ work-smart-backend/aircore/src/main/java/com/aircore/repository/IpRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.aircore.entity.ValidateIP;
 import com.aircore.utility.Enumeration.Status;
 
-public interface IpRepository extends JpaRepository<ValidateIP, Long>{
+public interface IpRepository extends JpaRepository<ValidateIP, Integer>{
 
 	ValidateIP findByIpAddressAndStatus(String trim, Status active);
